Don't write messages without (partial) body to database
Opening such messages during download will display "No text" and (probably
due to a bug) might lead to the synchronization process being aborted. Instead
of fixing the UI issue we now don't write these incomplete messages to the
database. This has the potential to massively speed up the sync process. But
it will take longer for messages to show up in the message list, especially
with slow connections.