IMAP keywords: a keyword->tag mapping activity
This creates a very simple activity (shamelessly copying from FolderList) to
present a listing of imap keywords and their corresponding tags from the memory
map. Read-only, for now.

To access the activity, click the Tags label (or any of the items) in the
message header.

Also, change Flag.REMEMBERED_KEYWORDS to LinkedHashMap to maintain
iteration order.