dynamically check if linkify URI schemes are supported by any apps
In order to prevent pointlessly confusing linkifying in the message body,
K-9 should only linkify special links (i.e. non-HTTP) if there is an app
installed that can handle that URI scheme.  This happens in onResume() in
both MessageList and MessageCompose so that the info is reasonably current
when the linkifying happens.