Created stub files for the widget.
Removed some log messages.
Created a xml file for an individual email item. Added the configuration activity on the AndroidManifest file.

First working demo.

Created a MailMessageDTO that will hold the data.
An ArrayList of MailMessageDTO holds the data to be displayed

Stylistic changes.
Added time to each email item. Also changed the font and the layout of
each email item.`

for wip pull request!

added intent to launch MessageList activity when the user taps in a mail item

Delete widget configuration when removing the widget.

Update the list view when a new email arrives.
Or one is deleted or moved

added intent to launch MessageList activity when the user taps in a mail item

UI changes.
Change the color of a mail item based on it's unread property.
Removed sender address from projection list

added padding in the main layout of the widget
and changed text size on subject and preview text

Added button to compose email.

Add custom action strings.

Replace Intent.ACTION_VIEW with custom action strings.

Display an attachement icon if the mail has attachment.

Changes on the ui.

Add widget icon.