Move class to feature package
Removed m prefix from variables.

Throw the exception if we are building in debug mode.

Removed comments, put MaiListRemoteViewFactory class in a separate file.

Set text size directly on layout file.

temp commit.

Changes on the widget layout.
Removed trivial colors, and fixed values on strings.xml

Changes on the ui functionality.
Added textview that shows the email of the inbox.
The color of the text of an unread email is now black and changes to
grey if it becomes read.

Use Collection#clear() before querying content provider.

add a loading view