UTC time in quotes (Issue #2403)
Before, the time in the quoted message of a reply was always in the
local timezone, even when the "Hide timezone" setting was selected. Now,
when the option is enabled, that is changed to UTC just like the time in
headers. I also changed the description for the option in the English
strings.xml to better describe the new different behavior.