For Android > N handle potential FileUriExposedException (see desc.)
Starting in Nougat, Android wants you to use a content:// URI and throws a
FileUriexposedException if your intent passes a file:// to another app.

The correct, full solution is to impliment Storage Access Framework across
k-9.  But until that happens, we'll catch the exception and use the
fallback dialog to ask for a path directly.