Use sensible shortcut keys (fixes #3057)  (ANDROID STUDIO 3.1)
See Issue #3057 for discussion, but this creates shortcut keys that are
intuitive in a desktop (chromebook) or hard keyboard setting.  Aside from
using traditional ctrl-modifiers to take important actions, the keys are
chosen to not clobber over native Android EditText behavior such as copy,
paste, select, navigate, undo/redo, etc.  The specific changes are in the
strings.xml.  Of note, the ctrl-RETURN to send doesn't function on my
chromebook, but I did do it according to the documentation, so it's very
possible this is a an issue with the framework itself.

The modifier key support required compat views, which in turn required
more recent APIs, which in turn required the latest Android Studio stuff.
So this commit is built over my previous compatibility changes to the
Android 3.1 build environment.  (PR #3159, etc.)