Remove Google from author field in About screen
K-9 Mail started off as a fork of the AOSP Email app. But this was
nearly 10 years ago. Most of the app has been rewritten or significantly
changed since then. By now I think it's fair to only list the K-9
project as author in the About screen.
Copyright statements in license headers and the copyright statement in
the About screen remain unaffected by this change.