Fix build problems with a fresh forked k-9
At the moment, when you checkout K-9 from scratch and install the latest
Android Studio, there are severeral problems with `gradle build`:

- removed uses-sdk from AndroidManifest.xml
- updated buildTools to 28.0.3
- updated kotlin to 1.2.71
- set Gradle jvmargs to prevent from OutOfMemory
- set targetSdk to 26 as AS sees all <26 as an Error because of
  Google Play.
- upgrade to latest Gradle 4.x version

I also excluded 2 Files from packaging as they are created more than
twice. Don't know why this happens and it should be fixed sometime but
as the files are empty I hope this is ok.