Migrate to AndroidX.
Here are the changes that fix build problems from the previous
commit that included all changes from the automatically migration.

Notice that some libraries are forced to alpha-versions at the moment.
They are introduced somehow through the automatically migration and
I didn't wanted to step back (this is not so easy and I don't know if
it will work.)