Incremented Settings.VERSION to define new version of settings.
The version integral needs to be incremented according to comment in GlobalSettings.java:47.
This new version number is applied to the newly added setting "messageViewWithoutHtml".

This commit is prone to merge conflicts. To resolve such a conflict increment the version from their branch and apply new number on all places changed within this commit.