Add detection of configuration based on SRV records (RFC6186)
If "XML Discovery" yields no connection settings, query SRV records
on the domain portion of the email address. This will help increase
the likelihood of being able to set up accounts with only the email and
password.

This implementation prefers TLS incoming servers (IMAPs/POP3s) with the
lowest priority value. If neither can be found, it will fall back to
using IMAP/POP3 with STARTTLS security. It is then left to
finishAutoSetup() to handle testing of the connection.

Things this commit hasn't covered:

  * Domain validation ("the user should be alerted if the SRV target
      is not part of the email domain")

  * Setting SSL/TLS security for the outbound server (it assumes
      STARTTLS currently)

NB: I attempted to merge the code that @daquexian had worked on but it
has drifted so far behind master that the amount of effort to merge that
in would be very large.

References https://github.com/k9mail/k-9/issues/2530