Extend the Account class to mute specific senders
To avoid notification hell, it is sometimes desirable to have a way to
_not_ get notified about mails originating from specific senders.

This change introduces support for suppressing notifications from a
given set of senders.

Note: this commit does not yet introduce any user interface to specify
those senders. This is conceptually different from the logic introduced
here, and is left for a later commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>