Optionally suppress notifications for mails sent to given recipients
On some mailing lists, policy requires that all replies be sent via
reply-to-all. This means that recipients subscribed to said mailing
lists might receive those replies twice, once with the `List-ID` header,
and once without. In these instances, the just-introduced option to
suppress notifications for mails sent via a mailing list won't catch the
second such reply.

To help with that, let's offer to suppress notifications for mails for
given recipients; That way, the mailing list address can be specified in
order to suppress also those second replies.

Note: the UI for this will follow in a later commit, to make reviewing
the Pull Request easier.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>