This works around a tiny issue made worse with the unified inbox as folder presentation.
If you disable the unified inbox while still viewing it, you could end
up with no account "selected", in a drawer showing folders.  This could
lead to a crash if the user clicks on a folder before selecting a
profile.  This patch resets the drawer to the default account, and
deselects any folder in the drawer.