Don't start another instance of AccountSettingsActivity when removing an account
Removing an account will trigger an update of AccountSelectionSpinner which will then detect the currently selected account is no longer available. It'll mark the default account as selected which in turn will trigger a call to onAccountSelected(). That will start AccountSettingsActivity with the newly selected account.
We now check whether finish() has already been called (because deleting the account will do that) and not start another instance of AccountSettingsActivity in that case.