Don't reuse AccountsLiveData instance
When the last account is deleted while AccountsLiveData is active it will hold on to an empty list. Then, when an account is added again and AccountsLiveData becomes active again, the empty list is emitted before the list containing the new account is emitted. This lead to the onboarding screen being shown when it shouldn't have. Not holding on to old AccountsLiveData instances will get rid of this problem.