Remove direct share support
We (can) no longer use the contacts database to keep track on how often someone has been contacted. Without that information we can't make meaningful automatic suggestions for share targets.
In the future we want to add support for app shortcuts to create a new message to a specific contact (see issue #2145).