Ignore WebView's scroll state when swipe gesture started in other view
We only consider a gesture as swipe gesture for moving to the previous or next message when the WebView can't be scrolled further in that direction. This is the correct behavior when the swipe gesture was started inside the WebView. However, the WebView's scroll state should be ignored when the swipe gesture was started e.g. in the message header. In those cases we always want to swipe to the next/previous message.