Message Details: Update list when the fragment is resumed
We want to update the list of participants when the associated contact has changed in the contacts database. Right now we don't listen to changes to the contacts database, but refresh the list when onResume() is called. This should catch the case where a user is adding a contact from the message details screen using one of the "add to contacts" buttons.