Fix F-Droid build
fdroidserver contains code to strip signing config blocks from `build.gradle[.kts]` files. That code also removed the assignment inside the `let` lambda. This in turn lead to Gradle failing the compilation of the Kotlin script because the argument `releaseSigningConfig` was unused 😞

Adding the "F-Droid hack" comment in this line prevents fdroidserver's regular expression from matching and removing this line.