Change `StoreSchemaDefinitionTest` to test tables for compatibility
Sometimes we want to remove columns from a database table. However, not all Android versions ship with a SQLite version that supports dropping columns. So, compared to a newly created table, a migrated table might contain additional (abandoned, but not removed) columns. This commit changes the test to check if table schemas have been migrated correctly to allow for additional columns in upgraded tables.