Add (or rather declare) support for RFC 6531, the SMTPUTF8 extension.
This gives Thunderbird the ability to send messages to/from unicode
addresses such as grå@grå.org. The user interface is not changed
by this; it will be changed in a subsequent commit.

Thunderbird already supported the addresses substantively, since it uses
the String type for addresses. That class has no support for ASCII-only
strings, only for unicode strings. In effect, the programming language
prevented Thunderbird from being incompatible with unicode addresses.

RFC 6531 requires more than substantive compatibility though, it also
requires correct declaration of the addresses in the SMTP protocol, and
that's what this commit adds.