Repair addresses that arrive with a punycode-damaged domain.
An header such as
  From: info@xn--gr-zia.org
is syntax according to RFCs 822 and 5322, but should not be shown to users
because the "xn--gr-zia" is as user-unfriendly as it is phisher-friendly.
This commit makes Thunderbird morph addresses that used xn-- into the pure
unicode form before they can be shown to the user.

This also implies that the human-readable form is used in replies.