Repair addresses that arrive with a punycode-damaged domain.
An header such as
  From: info@xn--gr-zia.org
is syntax according to RFCs 822 and 5322, but should not be shown to users
because the "xn--gr-zia" is as user-unfriendly as it is phisher-friendly.
This commit makes Thunderbird morph addresses that used xn-- into the pure
unicode form on input.

This change is not strictly related to EAI. However, it depends on
SMTPUTF8 to keep the code simple: When a user replies to a message that
used punycode, the reply's syntax often depends on RFC 6532, and sending
the reply needs RFC 6531.