Add merge scripts
The merge scripts help with merge automation. They need to be run
by a human at this point to ensure any conflicts can be handled
manually.

The scripts were used with the 11.0b1 merge of main->beta,
and have not yet been used on a merge of beta->release. They will
likely need further updates after the next merge of beta->release
is performed.